requirements.txt
amqp==5.1.0
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==3.2.5
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==3.3
kombu==5.2.4
mysqlclient==2.0.3
prompt-toolkit==3.0.29
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
